requirements.txt
mysql-connector-python
python-dotenv
debugpy
black